revamped database calls, database structure, added missing images
Updates:

GENERAL

The access to the database has been reworked and now it uses two config
files. Will post generic samples of these in the next revision. All
working files have been updated to reflect this change.

--

For lista.php we also cleaned the ‘materias’ table, and adding a
‘materias & grupos’ relational table so that now:

* ‘Autores‘ now show what semester and study program they are on in
parenthesis, right after their name. This takes the study program they
are on from the ‘alumno‘ table and joins it with the appropriate
semester grade that course is offered for their study program.
* the downloads structure is cleaner/leaner now

---

For the form (index.php)

* added a more detailed ‘especificaciones’ input section and set it up
in tabs.
* added an autocomplete function for adding names of ‘autores’
* added a new ‘descripción’ form element, so that students may include
a detailed description of their project